refactor: unify cloud database entities with enterprise

+++ packages/server/api/src/app/database/migration/postgres/1701084418793-add-state-to-otp.ts
@@ -1,10 +1,15 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../database-common'
 import { logger } from '@activepieces/server-shared'
+import { ApEdition } from '@activepieces/shared'
 
 export class AddStateToOtp1701084418793 implements MigrationInterface {
     name = 'AddStateToOtp1701084418793'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "otp"
             ADD "state" character varying NOT NULL
@@ -14,6 +19,9 @@ export class AddStateToOtp1701084418793 implements MigrationInterface {
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "otp" DROP COLUMN "state"
         `)
