+++ packages/server/api/src/app/database/migration/postgres/1705969874745-MakePlatformNotNullable.ts
@@ -1,9 +1,14 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../database-common'
+import { ApEdition } from '@activepieces/shared'
 
 export class MakePlatformNotNullable1705969874745 implements MigrationInterface {
     name = 'MakePlatformNotNullable1705969874745'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "project" DROP COLUMN "type"
         `)
@@ -25,6 +30,9 @@ export class MakePlatformNotNullable1705969874745 implements MigrationInterface 
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "project" DROP CONSTRAINT "fk_project_platform_id"
         `)
