+++ packages/server/api/src/app/database/migration/postgres/1709753080714-AddMappingStateToGit.ts
@@ -1,9 +1,14 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../database-common'
+import { ApEdition } from '@activepieces/shared'
 
 export class AddMappingStateToGit1709753080714 implements MigrationInterface {
     name = 'AddMappingStateToGit1709753080714'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "git_repo"
             ADD "mapping" jsonb
@@ -11,6 +16,9 @@ export class AddMappingStateToGit1709753080714 implements MigrationInterface {
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(`
             ALTER TABLE "git_repo" DROP COLUMN "mapping"
         `)
